server/pyproject.toml
[tool.poetry]
name = "hray"
version = "0.1.0"
description = "Hray server"
authors = ["Di Nguyen <hdinguyen@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
dspy = "2.6.0rc6"
fastapi = "0.111.1"
uvicorn = "0.22.0"
python-dotenv = "^1.0.1"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-exporter-jaeger = "1.21.0"
opentelemetry-instrumentation = "0.42b0"
opentelemetry-exporter-otlp = "^1.21.0"
opentelemetry-instrumentation-fastapi = "0.42b0"
opentelemetry-semantic-conventions = "0.42b0"
openinference-instrumentation-dspy = "^0.1.0"
arize-phoenix = "^7.3.2"
arize-phoenix-otel = "^0.6.1"
openinference-instrumentation-litellm = "^0.1.0"
openai = ">=1.55.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
